[#169] fix types homeless-people-help-map
