+++ apps/homeless-people-help-map/src/utils/types.ts
@@ -1,11 +1,19 @@
 import { FeatureCollection } from "@bratislava/utils/src/types";
+import { z } from "zod";
 
-export interface ITerrainService {
-  key: string;
-  title: string;
-  provider: string;
-  phone: string;
-  price: string;
-  areas: string;
+export const terrainServicePropertiesSchema = z.object({
+  key: z.string(),
+  title: z.string().optional(),
+  provider: z.string().optional(),
+  phone: z.number().optional().nullable(),
+  // web: z.string(),
+  // openingHours: z.string(),
+  price: z.string().optional().nullable(),
+  areas: z.string().optional(),
+});
+
+export type TerrainServiceProperties = z.infer<typeof terrainServicePropertiesSchema>;
+
+export interface ITerrainService extends TerrainServiceProperties {
   geojson: FeatureCollection;
 }
